refactor(tui): extract overlay and layout state modules
Move the mode/command-palette/picker model into state/overlay.rs (Mode,
Command, LeaderAction, PaletteState, PickerState, parse_leader,
leader_action) and the detail-viewport geometry + mouse hit-testing into
state/layout.rs (Split, DetailLayout, ClickTarget, ClickZone). Both reached
by qualified path; DetailLayout/ClickZone methods and leader_action the
reducer calls are pub(super). ui.rs and the outer tui/mod.rs import from
state::overlay / state::layout. No behavior change.
